requirements.txt
alembic==1.12.0
asgiref==3.7.2
blinker==1.6.3
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
filelock==3.12.4
greenlet==3.0.0
itsdangerous==2.1.2
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==68.2.2
SQLAlchemy==2.0.22
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.6
Werkzeug==3.0.0
